requirement.txt
Flask==2.2.5
flask-socketio==5.3.2
eventlet==0.33.3
transformers==4.35.0
torch==2.2.0
scikit-learn==1.3.0
pymongo==4.3.3
mimesis==18.0.0
python-dotenv==1.0.0
